checking lat long difference

+++ NaptanDataPrototype/Models/NaptanModel.cs
@@ -2,8 +2,8 @@ namespace NaptanDataPrototype.Models;
 
 public class NaptanModel
 {
-    public double Easting { get; set; }
-    public double Northing { get; set; }
+    public int Easting { get; set; }
+    public int Northing { get; set; }
     public double Latitude { get; set; }
     public double Longitude { get; set; }
 }
